+++ src/test/scala/model/MoneyTest.scala
@@ -14,7 +14,7 @@ class MoneyTest extends WordSpec
   "Money" when {
     "added to other Money" must {
       "add the underlying amount" in {
-        forAll { (amount1: Double, amount2: Double) =>
+        forAll { (amount1: BigDecimal, amount2: BigDecimal) =>
           Money(amount1) + Money(amount2) mustBe Money(amount1 + amount2)
         }
       }
@@ -22,7 +22,7 @@ class MoneyTest extends WordSpec
 
     "subtracted from other Money" must {
       "subtract the underlying amount" in {
-        forAll { (amount1: Double, amount2: Double) =>
+        forAll { (amount1: BigDecimal, amount2: BigDecimal) =>
           Money(amount1) - Money(amount2) mustBe Money(amount1 - amount2)
         }
       }
@@ -30,7 +30,7 @@ class MoneyTest extends WordSpec
 
     "multiplied by an integer" must {
       "multiply the underlying amount" in {
-        forAll { (amount: Double, multiple: Int) =>
+        forAll { (amount: BigDecimal, multiple: Int) =>
           Money(amount) * multiple  mustBe Money(amount * multiple)
         }
       }
